Separa en un componente el card de la noticia
-Separa el card de la noticia en un componente
reutilizable (noticia_card.dart).
-Agrega manejo de error en la imagen traida desde la API.
